requirements.txt
numpy==1.24.1
qiskit-aer==0.12.0
qiskit-algorithms==0.2.2
qiskit-ibmq-provider==0.20.2
qiskit-nature[pyscf]==0.7.1
qiskit-terra==0.45.2
scipy==1.9.1